Moving MiningData into a shared type
