+++ web/components/RealtimeTable.tsx
@@ -11,29 +11,12 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { Transaction, address, networks } from "bitcoinjs-lib";
+import { MiningData } from "@/lib/types";
 
 /* -----------------------------------
    Type Definitions
 ----------------------------------- */
-export interface MiningData {
-  pool_name: string;
-  timestamp: string;
-  height: number;
-  prev_hash: string;    // raw previous block hash (hex)
-  version: string;      // version (e.g. "20000000")
-  coinbase1: string;
-  coinbase2: string;
-  clean_jobs: boolean | string;
-  first_transaction: string; // computed from merkle_branches
-  fee_rate: number | string; // possibly empty on arrival
-  merkle_branches: string[];
-  merkle_branch_colors?: string[];
-  extranonce1: string;
-  extranonce2_length: number;
-  coinbase_outputs?: { address: string; value: number }[];
-  nbits?: string;
-  ntime?: string;
-}
+// MiningData interface is now imported from @/lib/types
 
 // Internal "enhanced" row with derived fields
 interface SortedRow extends MiningData {
